Clean up props
